refactor(core): rename type.ts to types.ts in Button and Toast for consistency
