+++ packages/core/src/components/Toast/index.tsx
@@ -4,7 +4,7 @@ import { notifications } from '@mantine/notifications';
 
 import styles from './styles.module.scss';
 
-import type { IToastOptions, TToastInput } from './type';
+import type { IToastOptions, TToastInput } from './types';
 
 interface IToastRuntime {
   (input: TToastInput): void;
